agent/requirements.txt
uvicorn==0.34.3
fastapi==0.115.13
python-dotenv==1.1.0
httpx==0.28.1